The big break-out/refactor
